extract set up logic from main
